Database: Selection refactored to BaseResultSet
