+++ Nette/Database/Table/Selection.php
@@ -12,7 +12,9 @@
 namespace Nette\Database\Table;
 
 use Nette,
-	Nette\Database\ISupplementalDriver;
+	Nette\Caching\IStorage,
+	Nette\Database\Connection,
+	Nette\Database\IReflection;
 
 
 /**
@@ -21,46 +23,15 @@ use Nette,
  *
  * @author     Jakub Vrana
  * @author     Jan Skrasek
- *
- * @property-read string $sql
  */
-class Selection extends Nette\Object implements \Iterator, IRowContainer, \ArrayAccess, \Countable
+class Selection extends BaseResultSet implements \ArrayAccess
 {
-	/** @var Nette\Database\Connection */
-	protected $connection;
-
-	/** @var Nette\Database\IReflection */
-	protected $reflection;
-
-	/** @var Nette\Caching\Cache */
-	protected $cache;
 
 	/** @var SqlBuilder */
 	protected $sqlBuilder;
 
-	/** @var string table name */
-	protected $name;
-
-	/** @var string primary key field name */
-	protected $primary;
-
-	/** @var string|bool primary column sequence name, FALSE for autodetection */
-	protected $primarySequence = FALSE;
-
-	/** @var IRow[] data read from database in [primary key => IRow] format */
-	protected $rows;
-
-	/** @var IRow[] modifiable data in [primary key => IRow] format */
-	protected $data;
-
 	/** @var bool */
 	protected $dataRefreshed = FALSE;
-
-	/** @var mixed cache array of Selection and GroupedSelection prototypes */
-	protected $globalRefCache;
-
-	/** @var mixed */
-	protected $refCache;
 
 	/** @var string */
 	protected $generalCacheKey;
@@ -80,22 +51,17 @@ class Selection extends Nette\Object implements \Iterator, IRowContainer, \Array
 	/** @var bool should instance observe accessed columns caching */
 	protected $observeCache = FALSE;
 
-	/** @var array of primary key values */
-	protected $keys = array();
-
 
 	/**
 	 * Creates filtered table representation.
-	 * @param  Nette\Database\Connection
-	 * @param  string  database table name
+	 * @param  Connection
+	 * @param  string
+	 * @param  IReflection
+	 * @param  IStorage
 	 */
-	public function __construct(Nette\Database\Connection $connection, $table, Nette\Database\IReflection $reflection, Nette\Caching\IStorage $cacheStorage = NULL)
+	public function __construct(Connection $connection, $table, IReflection $reflection, IStorage $cacheStorage = NULL)
 	{
-		$this->name = $table;
-		$this->connection = $connection;
-		$this->reflection = $reflection;
-		$this->cache = $cacheStorage ? new Nette\Caching\Cache($cacheStorage, 'Nette.Database.' . md5($connection->getDsn())) : NULL;
-		$this->primary = $reflection->getPrimary($table);
+		parent::__construct($table, $connection, $reflection, $cacheStorage);
 		$this->sqlBuilder = new SqlBuilder($table, $connection, $reflection);
 		$this->refCache = & $this->getRefTable($refPath)->globalRefCache[$refPath];
 	}
@@ -110,79 +76,6 @@ class Selection extends Nette\Object implements \Iterator, IRowContainer, \Array
 	public function __clone()
 	{
 		$this->sqlBuilder = clone $this->sqlBuilder;
-	}
-
-
-	/**
-	 * @return Nette\Database\Connection
-	 */
-	public function getConnection()
-	{
-		return $this->connection;
-	}
-
-
-	/**
-	 * @return Nette\Database\IReflection
-	 */
-	public function getDatabaseReflection()
-	{
-		return $this->reflection;
-	}
-
-
-	/**
-	 * @return string
-	 */
-	public function getName()
-	{
-		return $this->name;
-	}
-
-
-	/**
-	 * @param  bool
-	 * @return string|array
-	 */
-	public function getPrimary($need = TRUE)
-	{
-		if ($this->primary === NULL && $need) {
-			throw new \LogicException("Table \"{$this->name}\" does not have a primary key.");
-		}
-		return $this->primary;
-	}
-
-
-	/**
-	 * @return string
-	 */
-	public function getPrimarySequence()
-	{
-		if ($this->primarySequence === FALSE) {
-			$this->primarySequence = NULL;
-			$driver = $this->connection->getSupplementalDriver();
-			if ($driver->isSupported(ISupplementalDriver::SUPPORT_SEQUENCE) && $this->primary !== NULL) {
-				foreach ($driver->getColumns($this->name) as $column) {
-					if ($column['name'] === $this->primary) {
-						$this->primarySequence = $column['vendor']['sequence'];
-						break;
-					}
-				}
-			}
-		}
-
-		return $this->primarySequence;
-	}
-
-
-	/**
-	 * @param  string
-	 * @return self
-	 */
-	public function setPrimarySequence($sequence)
-	{
-		$this->primarySequence = $sequence;
-		return $this;
 	}
 
 
@@ -237,35 +130,6 @@ class Selection extends Nette\Object implements \Iterator, IRowContainer, \Array
 		return $clone->wherePrimary($key)->fetch();
 	}
 
-
-	/**
-	 * @inheritDoc
-	 */
-	public function fetch()
-	{
-		$this->execute();
-		$return = current($this->data);
-		next($this->data);
-		return $return;
-	}
-
-
-	/**
-	 * @inheritDoc
-	 */
-	public function fetchPairs($key = NULL, $value = NULL)
-	{
-		return Nette\Database\Helpers::toPairs(iterator_to_array($this), $key, $value);
-	}
-
-
-	/**
-	 * @inheritDoc
-	 */
-	public function fetchAll()
-	{
-		return iterator_to_array($this);
-	}
 
 
 	/********************* sql selectors ****************d*g**/
@@ -430,8 +294,7 @@ class Selection extends Nette\Object implements \Iterator, IRowContainer, \Array
 	public function count($column = NULL)
 	{
 		if (!$column) {
-			$this->execute();
-			return count($this->data);
+			return parent::count();
 		}
 		return $this->aggregation("COUNT($column)");
 	}
@@ -521,17 +384,6 @@ class Selection extends Nette\Object implements \Iterator, IRowContainer, \Array
 		return new ActiveRow($row, $this);
 	}
 
-
-	public function createSelectionInstance($table = NULL)
-	{
-		return new Selection($this->connection, $table ?: $this->name, $this->reflection, $this->cache ? $this->cache->getStorage() : NULL);
-	}
-
-
-	protected function createGroupedSelectionInstance($table, $column)
-	{
-		return new GroupedSelection($table, $column, $this, $this->connection, $this->reflection, $this->cache ? $this->cache->getStorage() : NULL);
-	}
 
 
 	protected function query($query)
@@ -778,108 +630,6 @@ class Selection extends Nette\Object implements \Iterator, IRowContainer, \Array
 		return $this->query($this->sqlBuilder->buildDeleteQuery())->getRowCount();
 	}
 
-
-	/********************* references ****************d*g**/
-
-
-	/**
-	 * Returns referenced row.
-	 * @param  string
-	 * @param  string
-	 * @param  mixed   primary key to check for $table and $column references
-	 * @return Selection or array() if the row does not exist
-	 */
-	public function getReferencedTable($table, $column, $checkPrimaryKey)
-	{
-		$referenced = & $this->refCache['referenced'][$this->getSpecificCacheKey()]["$table.$column"];
-		$selection = & $referenced['selection'];
-		$cacheKeys = & $referenced['cacheKeys'];
-		if ($selection === NULL || !isset($cacheKeys[$checkPrimaryKey])) {
-			$this->execute();
-			$cacheKeys = array();
-			foreach ($this->rows as $row) {
-				if ($row[$column] === NULL) {
-					continue;
-				}
-
-				$key = $row[$column];
-				$cacheKeys[$key] = TRUE;
-			}
-
-			if ($cacheKeys) {
-				$selection = $this->createSelectionInstance($table);
-				$selection->where($selection->getPrimary(), array_keys($cacheKeys));
-			} else {
-				$selection = array();
-			}
-		}
-
-		return $selection;
-	}
-
-
-	/**
-	 * Returns referencing rows.
-	 * @param  string
-	 * @param  string
-	 * @param  int primary key
-	 * @return GroupedSelection
-	 */
-	public function getReferencingTable($table, $column, $active = NULL)
-	{
-		$prototype = & $this->refCache['referencingPrototype']["$table.$column"];
-		if (!$prototype) {
-			$prototype = $this->createGroupedSelectionInstance($table, $column);
-			$prototype->where("$table.$column", array_keys((array) $this->rows));
-		}
-
-		$clone = clone $prototype;
-		$clone->setActive($active);
-		return $clone;
-	}
-
-
-	/********************* interface Iterator ****************d*g**/
-
-
-	public function rewind()
-	{
-		$this->execute();
-		$this->keys = array_keys($this->data);
-		reset($this->keys);
-	}
-
-
-	/** @return IRow */
-	public function current()
-	{
-		if (($key = current($this->keys)) !== FALSE) {
-			return $this->data[$key];
-		} else {
-			return FALSE;
-		}
-	}
-
-
-	/**
-	 * @return string row ID
-	 */
-	public function key()
-	{
-		return current($this->keys);
-	}
-
-
-	public function next()
-	{
-		next($this->keys);
-	}
-
-
-	public function valid()
-	{
-		return current($this->keys) !== FALSE;
-	}
 
 
 	/********************* interface ArrayAccess ****************d*g**/
